+++ modules/katamari/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The `Result` type is now covariant with respect to its type argument.
 - All singletons now have a `get` function that returns an `Optional` value.
 - Made constructing `Optional`s faster by re-implementing them as a class.
+- Made constructing `Result`s faster by re-implementing them as a class.
 
 ### Fixed
 - `Throttler.adaptable` could not re-throttle from within the callback function.